Move LocalIO into its own submodule
