+++ consensus/src/member.rs
@@ -1,94 +1,14 @@
 use crate::{
     dissemination::{Addressed, DisseminationMessage},
     handle_task_termination,
+    interface::LocalIO,
     network::{Hub as NetworkHub, NetworkData, UnitMessage},
     runway::{self, NetworkIO, RunwayIO},
-    Config, Data, DataProvider, FinalizationHandler, Hasher, MultiKeychain, Network, NodeIndex,
-    OrderedUnit, Receiver, Recipient, Sender, Signature, SpawnHandle, Terminator,
-    UnitFinalizationHandler,
+    Config, Data, DataProvider, Hasher, MultiKeychain, Network, NodeIndex, Receiver, Recipient,
+    Sender, Signature, SpawnHandle, Terminator, UnitFinalizationHandler,
 };
 use futures::{channel::mpsc, pin_mut, AsyncRead, AsyncWrite, FutureExt, StreamExt};
 use log::{debug, error, info, warn};
-use std::marker::PhantomData;
-
-/// This adapter allows to map an implementation of [`FinalizationHandler`] onto implementation of [`UnitFinalizationHandler`].
-pub struct FinalizationHandlerAdapter<FH, D, H> {
-    finalization_handler: FH,
-    _phantom: PhantomData<(D, H)>,
-}
-
-impl<FH, D, H> From<FH> for FinalizationHandlerAdapter<FH, D, H> {
-    fn from(value: FH) -> Self {
-        Self {
-            finalization_handler: value,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<D: Data, H: Hasher, FH: FinalizationHandler<D>> UnitFinalizationHandler
-    for FinalizationHandlerAdapter<FH, D, H>
-{
-    type Data = D;
-    type Hasher = H;
-
-    fn batch_finalized(&mut self, batch: Vec<OrderedUnit<Self::Data, Self::Hasher>>) {
-        for unit in batch {
-            if let Some(data) = unit.data {
-                self.finalization_handler.data_finalized(data)
-            }
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct LocalIO<DP: DataProvider, UFH: UnitFinalizationHandler, US: AsyncWrite, UL: AsyncRead> {
-    data_provider: DP,
-    finalization_handler: UFH,
-    unit_saver: US,
-    unit_loader: UL,
-}
-
-impl<
-        H: Hasher,
-        DP: DataProvider,
-        FH: FinalizationHandler<DP::Output>,
-        US: AsyncWrite,
-        UL: AsyncRead,
-    > LocalIO<DP, FinalizationHandlerAdapter<FH, DP::Output, H>, US, UL>
-{
-    pub fn new(
-        data_provider: DP,
-        finalization_handler: FH,
-        unit_saver: US,
-        unit_loader: UL,
-    ) -> Self {
-        Self {
-            data_provider,
-            finalization_handler: finalization_handler.into(),
-            unit_saver,
-            unit_loader,
-        }
-    }
-}
-
-impl<DP: DataProvider, UFH: UnitFinalizationHandler, US: AsyncWrite, UL: AsyncRead>
-    LocalIO<DP, UFH, US, UL>
-{
-    pub fn new_with_unit_finalization_handler(
-        data_provider: DP,
-        finalization_handler: UFH,
-        unit_saver: US,
-        unit_loader: UL,
-    ) -> Self {
-        Self {
-            data_provider,
-            finalization_handler,
-            unit_saver,
-            unit_loader,
-        }
-    }
-}
 
 struct Member<H, D, S>
 where
@@ -203,14 +123,6 @@ where
 /// For a detailed description of the consensus implemented by `run_session` see
 /// [docs for devs](https://cardinal-cryptography.github.io/AlephBFT/index.html)
 /// or the [original paper](https://arxiv.org/abs/1908.05156).
-///
-/// Please note that in order to fulfill the constraint [`UnitFinalizationHandler<Data = DP::Output, Hasher
-/// = H>`] it is enough to provide implementation of [`FinalizationHandler<DP::Output>`]. We provide
-/// implementation of [`UnitFinalizationHandler<Data = DP::Output, Hasher = H>`] for anything that satisfies
-/// the trait [`FinalizationHandler<DP::Output>`] (by means of [`FinalizationHandlerAdapter`]). Implementing
-/// [`UnitFinalizationHandler`] directly is considered less stable since it exposes intrisics which might be
-/// subject to change. Implement [`FinalizationHandler<DP::Output>`] instead, unless you absolutely know
-/// what you are doing.
 pub async fn run_session<
     DP: DataProvider,
     UFH: UnitFinalizationHandler<Data = DP::Output>,
@@ -265,12 +177,8 @@ pub async fn run_session<
         unit_messages_from_network: runway_messages_from_network,
         unit_messages_for_network: runway_messages_for_network,
     };
-    let runway_io = RunwayIO::new(
-        local_io.data_provider,
-        local_io.finalization_handler,
-        local_io.unit_saver,
-        local_io.unit_loader,
-    );
+    let (data_provider, finalization_handler, unit_saver, unit_loader) = local_io.into_components();
+    let runway_io = RunwayIO::new(data_provider, finalization_handler, unit_saver, unit_loader);
     let spawn_copy = spawn_handle.clone();
     let runway_handle = spawn_handle
         .spawn_essential("member/runway", async move {
